🎨 Divide each structure & Introduce RunDockerError
